Created a new class uses a linked list of moves to track a reference to a page as it moves in Pack. Called MovablePosition.
Also created SkippableMovablePosition for when the next move is one that has already been recorded in Pack.

Created move implementations of MoveRecorder, one for the mutation and one for after the commit begins in Pack.

The new move tracking has created a MoveNode class in Pack.

Realizing that the goal post strategy is clever but almost redundant in Pack.



git-svn-id: http://blogometer.com/svn/open/trunk/pack@1988 3df9ba58-24f1-0310-ade5-ba95bd924a04